backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
gunicorn>=22.0.0
argon2-cffi>=23.1.0
authlib>=1.3.2
sentry-sdk[fastapi]>=2.19.0
python-jose[cryptography]==3.3.0
# Fernet (cifrado en reposo: IBAN, tokens bancarios, etc.)
cryptography>=42.0.0
httpx>=0.27.0
googlemaps>=4.10.0
passlib[argon2]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.3
pydantic-settings==2.1.0
supabase>=2.4.5
python-dotenv==1.0.0
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.0
redis>=5.0.0
limits[redis]>=3.13.0
slowapi>=0.1.9
azure-ai-formrecognizer>=3.3.0
azure-core>=1.29.0
fpdf2>=2.7.0
reportlab>=4.0.0
qrcode[pil]>=7.4.0
Pillow>=10.0.0
stripe>=11.0.0
gocardless-pro>=1.54.0
resend>=2.0.0
openai>=1.40.0
# SessionMiddleware (Starlette/FastAPI); firmado de cookies de sesión OAuth
itsdangerous>=2.1.2

pandas>=2.2.0
openpyxl>=3.1.0

# CI / tests
pytest>=8.0.0
pytest-asyncio>=0.24.0

# Pruebas de carga (Locust): no instalar en despliegue productivo — usar requirements-loadtest.txt
